+++ CHANGELOG.md
@@ -1,2 +1,2 @@
-- fixed 20.1 crash
+- fixed `minecraft:block` particles not being hidden in some scenarios
 - further improvements to overall system stability and other minor adjustments have been made to enhance the user experience